# carrieros-web/server/contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CarrierOS API",
    "version": "1.0.0",
    "description": "GENERATED from server/contract. Do not edit; run `npm run gen:api` in carrieros-web."
  },
  "paths": {
...
    "/api/v1/me/driver-profile": {
      "patch": {
        "operationId": "updateMyDriverProfile",
        "summary": "Update the caller's own driver record (CDL, endorsements, emergency contact, default vehicle). Protected fields are not accepted.",
        "tags": [
          "identity"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cdl_number": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 40
                  },
                  "cdl_class": {
                    "nullable": true,
                    "type": "string",
                    "enum": [
                      "A",
                      "B",
                      "C"
                    ]
                  },
                  "cdl_state": {
                    "nullable": true,
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 2
                  },
                  "endorsements": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "hazmat",
                        "tanker",
                        "doubles",
                        "airbrakes",
                        "passenger"
                      ]
                    }
                  },
                  "emergency_contact_name": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 100
                  },
                  "emergency_contact_phone": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 30
                  },
                  "emergency_contact_relation": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 50
                  },
                  "default_vehicle_id": {
                    "nullable": true,
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string",
                      "description": "Developer-facing fallback text; never render to end users."
                    },
                    "meta": {
                      "description": "Machine-readable context, e.g. { current_status } on a 409 VERSION_CONFLICT.",
                      "type": "object",
                      "additionalProperties": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "number"
                          },
                          {
                            "type": "boolean"
                          },
                          {
                            "type": "string",
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      }
                    }
                  },
                  "required": [
                    "error_code",
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string",
                      "description": "Developer-facing fallback text; never render to end users."
                    },
                    "meta": {
                      "description": "Machine-readable context, e.g. { current_status } on a 409 VERSION_CONFLICT.",
                      "type": "object",
                      "additionalProperties": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "number"
                          },
                          {
                            "type": "boolean"
                          },
                          {
                            "type": "string",
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      }
                    }
                  },
                  "required": [
                    "error_code",
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string",
                      "description": "Developer-facing fallback text; never render to end users."
                    },
                    "meta": {
                      "description": "Machine-readable context, e.g. { current_status } on a 409 VERSION_CONFLICT.",
                      "type": "object",
                      "additionalProperties": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "number"
                          },
                          {
                            "type": "boolean"
                          },
                          {
                            "type": "string",
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      }
                    }
                  },
                  "required": [
                    "error_code",
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string",
                      "description": "Developer-facing fallback text; never render to end users."
                    },
                    "meta": {
                      "description": "Machine-readable context, e.g. { current_status } on a 409 VERSION_CONFLICT.",
                      "type": "object",
                      "additionalProperties": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "number"
                          },
                          {
                            "type": "boolean"
                          },
                          {
                            "type": "string",
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      }
                    }
                  },
                  "required": [
                    "error_code",
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string",
                      "description": "Developer-facing fallback text; never render to end users."
                    },
                    "meta": {
                      "description": "Machine-readable context, e.g. { current_status } on a 409 VERSION_CONFLICT.",
                      "type": "object",
                      "additionalProperties": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "number"
                          },
                          {
                            "type": "boolean"
                          },
                          {
                            "type": "string",
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      }
                    }
                  },
                  "required": [
                    "error_code",
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}